+++ compiler-core/src/language_server/code_action.rs
@@ -21,7 +21,7 @@ use crate::{
     type_::{
         self, FieldMap, ModuleValueConstructor, Type, TypeVar, TypedCallArg, ValueConstructor,
         error::{ModuleSuggestion, VariableDeclaration, VariableOrigin},
-        printer::{Names, Printer},
+        printer::Printer,
     },
 };
 use ecow::{EcoString, eco_format};
@@ -1170,6 +1170,11 @@ impl<'ast> ast::visit::Visit<'ast> for AddAnnotations<'_> {
     }
 
     fn visit_typed_module_constant(&mut self, constant: &'ast TypedModuleConstant) {
+        // Since type variable names are local to definitions, any type variables
+        // in other parts of the module shouldn't affect what we print for the
+        // annotations of this constant.
+        self.printer.clear_type_variables();
+
         let code_action_range = self.edits.src_span_to_lsp_range(constant.location);
 
         // Only offer the code action if the cursor is over the statement
@@ -1189,6 +1194,14 @@ impl<'ast> ast::visit::Visit<'ast> for AddAnnotations<'_> {
     }
 
     fn visit_typed_function(&mut self, fun: &'ast ast::TypedFunction) {
+        // Since type variable names are local to definitions, any type variables
+        // in other parts of the module shouldn't affect what we print for the
+        // annotations of this functions. The only variables which cannot clash
+        // are ones defined in the signature of this function, which we register
+        // when we visit the parameters of this function inside `collect_type_variables`.
+        self.printer.clear_type_variables();
+        collect_type_variables(&mut self.printer, fun);
+
         ast::visit::visit_typed_function(self, fun);
 
         let code_action_range = self.edits.src_span_to_lsp_range(
@@ -1297,7 +1310,7 @@ impl<'a> AddAnnotations<'a> {
             edits: TextEdits::new(line_numbers),
             // We need to use the same printer for all the edits because otherwise
             // we could get duplicate type variable names.
-            printer: Printer::new(&module.ast.names),
+            printer: Printer::new_without_type_variables(&module.ast.names),
         }
     }
 
@@ -1318,6 +1331,24 @@ impl<'a> AddAnnotations<'a> {
             .changes(uri.clone(), self.edits.edits)
             .preferred(false)
             .push_to(actions);
+    }
+}
+
+struct TypeVariableCollector<'a, 'b> {
+    printer: &'a mut Printer<'b>,
+}
+
+/// Collect type variables defined within a function and register them for a
+/// `Printer`
+fn collect_type_variables(printer: &mut Printer<'_>, function: &ast::TypedFunction) {
+    TypeVariableCollector { printer }.visit_typed_function(function);
+}
+
+impl<'ast, 'a, 'b> ast::visit::Visit<'ast> for TypeVariableCollector<'a, 'b> {
+    fn visit_type_ast_var(&mut self, _location: &'ast SrcSpan, name: &'ast EcoString) {
+        // Register this type variable so that we don't duplicate names when
+        // adding annotations.
+        self.printer.register_type_variable(name.clone());
     }
 }
 
@@ -1392,47 +1423,6 @@ impl<'a> QualifiedToUnqualifiedImportFirstPass<'a> {
 }
 
 impl<'ast> ast::visit::Visit<'ast> for QualifiedToUnqualifiedImportFirstPass<'ast> {
-    fn visit_typed_expr_fn(
-        &mut self,
-        location: &'ast SrcSpan,
-        type_: &'ast Arc<Type>,
-        kind: &'ast FunctionLiteralKind,
-        arguments: &'ast [TypedArg],
-        body: &'ast Vec1<TypedStatement>,
-        return_annotation: &'ast Option<ast::TypeAst>,
-    ) {
-        for argument in arguments {
-            if let Some(annotation) = &argument.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-        if let Some(return_) = return_annotation {
-            self.visit_type_ast(return_);
-        }
-        ast::visit::visit_typed_expr_fn(
-            self,
-            location,
-            type_,
-            kind,
-            arguments,
-            body,
-            return_annotation,
-        );
-    }
-
-    fn visit_typed_function(&mut self, fun: &'ast ast::TypedFunction) {
-        for arg in &fun.arguments {
-            if let Some(annotation) = &arg.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-
-        if let Some(return_annotation) = &fun.return_annotation {
-            self.visit_type_ast(return_annotation);
-        }
-        ast::visit::visit_typed_function(self, fun);
-    }
-
     fn visit_type_ast_constructor(
         &mut self,
         location: &'ast SrcSpan,
@@ -1631,47 +1621,6 @@ impl<'a> QualifiedToUnqualifiedImportSecondPass<'a> {
 }
 
 impl<'ast> ast::visit::Visit<'ast> for QualifiedToUnqualifiedImportSecondPass<'ast> {
-    fn visit_typed_expr_fn(
-        &mut self,
-        location: &'ast SrcSpan,
-        type_: &'ast Arc<Type>,
-        kind: &'ast FunctionLiteralKind,
-        arguments: &'ast [TypedArg],
-        body: &'ast Vec1<TypedStatement>,
-        return_annotation: &'ast Option<ast::TypeAst>,
-    ) {
-        for argument in arguments {
-            if let Some(annotation) = &argument.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-        if let Some(return_) = return_annotation {
-            self.visit_type_ast(return_);
-        }
-        ast::visit::visit_typed_expr_fn(
-            self,
-            location,
-            type_,
-            kind,
-            arguments,
-            body,
-            return_annotation,
-        );
-    }
-
-    fn visit_typed_function(&mut self, fun: &'ast ast::TypedFunction) {
-        for arg in &fun.arguments {
-            if let Some(annotation) = &arg.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-
-        if let Some(return_annotation) = &fun.return_annotation {
-            self.visit_type_ast(return_annotation);
-        }
-        ast::visit::visit_typed_function(self, fun);
-    }
-
     fn visit_type_ast_constructor(
         &mut self,
         location: &'ast SrcSpan,
@@ -1876,46 +1825,6 @@ impl<'a> UnqualifiedToQualifiedImportFirstPass<'a> {
 }
 
 impl<'ast> ast::visit::Visit<'ast> for UnqualifiedToQualifiedImportFirstPass<'ast> {
-    fn visit_typed_expr_fn(
-        &mut self,
-        location: &'ast SrcSpan,
-        type_: &'ast Arc<Type>,
-        kind: &'ast FunctionLiteralKind,
-        arguments: &'ast [TypedArg],
-        body: &'ast Vec1<TypedStatement>,
-        return_annotation: &'ast Option<ast::TypeAst>,
-    ) {
-        for argument in arguments {
-            if let Some(annotation) = &argument.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-        if let Some(return_) = return_annotation {
-            self.visit_type_ast(return_);
-        }
-        ast::visit::visit_typed_expr_fn(
-            self,
-            location,
-            type_,
-            kind,
-            arguments,
-            body,
-            return_annotation,
-        );
-    }
-
-    fn visit_typed_function(&mut self, fun: &'ast ast::TypedFunction) {
-        for arg in &fun.arguments {
-            if let Some(annotation) = &arg.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-
-        if let Some(return_annotation) = &fun.return_annotation {
-            self.visit_type_ast(return_annotation);
-        }
-        ast::visit::visit_typed_function(self, fun);
-    }
     fn visit_type_ast_constructor(
         &mut self,
         location: &'ast SrcSpan,
@@ -2092,47 +2001,6 @@ impl<'a> UnqualifiedToQualifiedImportSecondPass<'a> {
 }
 
 impl<'ast> ast::visit::Visit<'ast> for UnqualifiedToQualifiedImportSecondPass<'ast> {
-    fn visit_typed_expr_fn(
-        &mut self,
-        location: &'ast SrcSpan,
-        type_: &'ast Arc<Type>,
-        kind: &'ast FunctionLiteralKind,
-        arguments: &'ast [TypedArg],
-        body: &'ast Vec1<TypedStatement>,
-        return_annotation: &'ast Option<ast::TypeAst>,
-    ) {
-        for argument in arguments {
-            if let Some(annotation) = &argument.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-        if let Some(return_) = return_annotation {
-            self.visit_type_ast(return_);
-        }
-        ast::visit::visit_typed_expr_fn(
-            self,
-            location,
-            type_,
-            kind,
-            arguments,
-            body,
-            return_annotation,
-        );
-    }
-
-    fn visit_typed_function(&mut self, fun: &'ast ast::TypedFunction) {
-        for arg in &fun.arguments {
-            if let Some(annotation) = &arg.annotation {
-                self.visit_type_ast(annotation);
-            }
-        }
-
-        if let Some(return_annotation) = &fun.return_annotation {
-            self.visit_type_ast(return_annotation);
-        }
-        ast::visit::visit_typed_function(self, fun);
-    }
-
     fn visit_type_ast_constructor(
         &mut self,
         location: &'ast SrcSpan,
@@ -3677,7 +3545,9 @@ impl<'a> GenerateDynamicDecoder<'a> {
         params: &'a CodeActionParams,
         actions: &'a mut Vec<CodeAction>,
     ) -> Self {
-        let printer = Printer::new(&module.ast.names);
+        // Since we are generating a new function, type variables from other
+        // functions and constants are irrelevant to the types we print.
+        let printer = Printer::new_without_type_variables(&module.ast.names);
         Self {
             module,
             params,
@@ -3765,7 +3635,7 @@ impl<'a> GenerateDynamicDecoder<'a> {
         }
 
         let mut decoder_printer = DecoderPrinter::new(
-            &self.module.ast.names,
+            &mut self.printer,
             custom_type.name.clone(),
             self.module.name.clone(),
         );
@@ -3873,8 +3743,8 @@ fn maybe_import(edits: &mut TextEdits<'_>, module: &Module, module_name: &str) {
     ));
 }
 
-struct DecoderPrinter<'a> {
-    printer: Printer<'a>,
+struct DecoderPrinter<'a, 'b> {
+    printer: &'a mut Printer<'b>,
     /// The name of the root type we are printing a decoder for
     type_name: EcoString,
     /// The module name of the root type we are printing a decoder for
@@ -3924,12 +3794,12 @@ impl RecordLabel<'_> {
     }
 }
 
-impl<'a> DecoderPrinter<'a> {
-    fn new(names: &'a Names, type_name: EcoString, type_module: EcoString) -> Self {
+impl<'a, 'b> DecoderPrinter<'a, 'b> {
+    fn new(printer: &'a mut Printer<'b>, type_name: EcoString, type_module: EcoString) -> Self {
         Self {
             type_name,
             type_module,
-            printer: Printer::new(names),
+            printer,
         }
     }
 
@@ -4069,7 +3939,9 @@ impl<'a> GenerateJsonEncoder<'a> {
         actions: &'a mut Vec<CodeAction>,
         config: &'a PackageConfig,
     ) -> Self {
-        let printer = Printer::new(&module.ast.names);
+        // Since we are generating a new function, type variables from other
+        // functions and constants are irrelevant to the types we print.
+        let printer = Printer::new_without_type_variables(&module.ast.names);
         Self {
             module,
             params,
@@ -4173,7 +4045,7 @@ impl<'a> GenerateJsonEncoder<'a> {
 
         // Otherwise we turn it into an object with a `type` tag field.
         let mut encoder_printer =
-            JsonEncoderPrinter::new(&self.module.ast.names, type_name, self.module.name.clone());
+            JsonEncoderPrinter::new(&mut self.printer, type_name, self.module.name.clone());
 
         // These are the fields of the json object to encode.
         let mut fields = Vec::with_capacity(constructor.arguments.len());
@@ -4258,20 +4130,20 @@ fn {name}({record_name}: {type_}) -> {json_type} {{
     }
 }
 
-struct JsonEncoderPrinter<'a> {
-    printer: Printer<'a>,
+struct JsonEncoderPrinter<'a, 'b> {
+    printer: &'a mut Printer<'b>,
     /// The name of the root type we are printing an encoder for
     type_name: EcoString,
     /// The module name of the root type we are printing an encoder for
     type_module: EcoString,
 }
 
-impl<'a> JsonEncoderPrinter<'a> {
-    fn new(names: &'a Names, type_name: EcoString, type_module: EcoString) -> Self {
+impl<'a, 'b> JsonEncoderPrinter<'a, 'b> {
+    fn new(printer: &'a mut Printer<'b>, type_name: EcoString, type_module: EcoString) -> Self {
         Self {
             type_name,
             type_module,
-            printer: Printer::new(names),
+            printer,
         }
     }
 
@@ -5114,7 +4986,10 @@ impl<'a> GenerateFunction<'a> {
         // generators to avoid renaming a label in case it shares a name with an argument.
         let mut label_names = NameGenerator::new();
         let mut argument_names = NameGenerator::new();
-        let mut printer = Printer::new(&self.module.ast.names);
+
+        // Since we are generating a new function, type variables from other
+        // functions and constants are irrelevant to the types we print.
+        let mut printer = Printer::new_without_type_variables(&self.module.ast.names);
         let arguments = arguments_types
             .iter()
             .enumerate()