+++ compiler-core/src/language_server/tests/action.rs
@@ -9224,3 +9224,81 @@ fn pattern_match_on_list_variable() {
         find_position_of("a_list").to_selection()
     );
 }
+
+// https://github.com/gleam-lang/gleam/issues/3786
+#[test]
+fn type_variables_from_other_functions_do_not_change_annotations() {
+    assert_code_action!(
+        ADD_ANNOTATIONS,
+        "
+fn wibble(a: a, b: b, c: c) -> d { todo }
+
+fn pair(a, b) {
+  #(a, b)
+}
+",
+        find_position_of("pair").to_selection()
+    );
+}
+
+#[test]
+fn type_variables_from_other_functions_do_not_change_annotations_constant() {
+    assert_code_action!(
+        ADD_ANNOTATION,
+        "
+fn wibble(a: a, b: b, c: c) -> d { todo }
+
+const empty = []
+",
+        find_position_of("empty").to_selection()
+    );
+}
+
+#[test]
+fn type_variables_are_not_duplicated_when_adding_annotations() {
+    assert_code_action!(
+        ADD_ANNOTATIONS,
+        "
+fn wibble(a: a, b: b, c: c) -> d { todo }
+
+fn many_args(a, b, c, d: d, e: a, f, g) {
+  todo
+}
+",
+        find_position_of("many_args").to_selection()
+    );
+}
+
+#[test]
+fn type_variables_in_let_bindings_are_considered_when_adding_annotations() {
+    assert_code_action!(
+        ADD_ANNOTATIONS,
+        "
+fn wibble(a, b, c) {
+  let x: a = todo
+  fn(a: b, b: c) -> d {
+    todo
+  }
+}
+",
+        find_position_of("wibble").to_selection()
+    );
+}
+
+#[test]
+fn generated_function_annotations_are_not_affected_by_other_functions() {
+    assert_code_action!(
+        GENERATE_FUNCTION,
+        "
+fn wibble(a: a, b: b, c: c) -> d { todo }
+
+pub fn main() {
+  let x = todo
+  let y = todo
+  let #(a, b) = something(x, y)
+  b
+}
+",
+        find_position_of("something").to_selection()
+    );
+}